+++ src/array/array_sync_writable.rs
@@ -4,14 +4,18 @@ use rayon::iter::{IndexedParallelIterator, IntoParallelIterator, ParallelIterato
 
 use crate::{
     array_subset::ArraySubset,
+    config::{MetadataOptionsEraseVersion, MetadataOptionsStoreVersion},
     metadata::array_metadata_v2_to_v3,
-    storage::{StorageError, StorageHandle, WritableStorageTraits},
+    storage::{
+        meta_key, meta_key_v2_array, meta_key_v2_attributes, StorageError, StorageHandle,
+        WritableStorageTraits,
+    },
 };
 
 use super::{
     codec::{options::CodecOptions, ArrayCodecTraits},
     concurrency::concurrency_chunks_and_codec,
-    Array, ArrayError, ArrayMetadata, ArrayMetadataOptions, ArrayMetadataOptionsVersion,
+    Array, ArrayError, ArrayMetadata, ArrayMetadataOptions,
 };
 
 impl<TStorage: ?Sized + WritableStorageTraits + 'static> Array<TStorage> {
@@ -27,12 +31,12 @@ impl<TStorage: ?Sized + WritableStorageTraits + 'static> Array<TStorage> {
         crate::storage::create_array(&*storage_transformer, self.path(), &self.metadata())
     }
 
-    /// Store metadata with non-default options.
+    /// Store metadata with non-default [`ArrayMetadataOptions`].
     ///
     /// # Errors
     /// Returns [`StorageError`] if there is an underlying store error.
     pub fn store_metadata_opt(&self, options: &ArrayMetadataOptions) -> Result<(), StorageError> {
-        use ArrayMetadataOptionsVersion as V;
+        use MetadataOptionsStoreVersion as V;
         let storage_handle = Arc::new(StorageHandle::new(self.storage.clone()));
         let storage_transformer = self
             .storage_transformers()
@@ -42,8 +46,8 @@ impl<TStorage: ?Sized + WritableStorageTraits + 'static> Array<TStorage> {
         let metadata = self.metadata_opt(options);
 
         // Convert/store the metadata as requested
-        match (metadata, options.array_metadata_version()) {
-            (ArrayMetadata::V3(metadata), V::Unchanged | V::V3) => {
+        match (metadata, options.metadata_store_version()) {
+            (ArrayMetadata::V3(metadata), V::Default | V::V3) => {
                 // Store V3
                 crate::storage::create_array(
                     &*storage_transformer,
@@ -53,17 +57,15 @@ impl<TStorage: ?Sized + WritableStorageTraits + 'static> Array<TStorage> {
             }
             (ArrayMetadata::V2(metadata), V::V3) => {
                 // Convert V2 to V3
-                let metadata = unsafe {
-                    // SAFETY: array_metadata_v2_to_v3 has already succeeded on array creation
-                    array_metadata_v2_to_v3(&metadata).unwrap_unchecked()
-                };
+                let metadata = array_metadata_v2_to_v3(&metadata)
+                    .map_err(|err| StorageError::from(err.to_string()))?;
                 crate::storage::create_array(
                     &*storage_transformer,
                     self.path(),
                     &ArrayMetadata::V3(metadata),
                 )
             }
-            (ArrayMetadata::V2(metadata), V::Unchanged) => {
+            (ArrayMetadata::V2(metadata), V::Default) => {
                 // Store V2
                 crate::storage::create_array(
                     &*storage_transformer,
@@ -72,7 +74,6 @@ impl<TStorage: ?Sized + WritableStorageTraits + 'static> Array<TStorage> {
                 )
             }
         }
-        // if let ArrayMetadataOptionsVersion::Unchanged = options.array_metadata_version() {}
     }
 
     /// Encode `chunk_bytes` and store at `chunk_indices`.
@@ -189,18 +190,46 @@ impl<TStorage: ?Sized + WritableStorageTraits + 'static> Array<TStorage> {
         self.store_chunks_ndarray_opt(chunks, chunks_array, &CodecOptions::default())
     }
 
-    /// Erase the metadata.
+    /// Erase the metadata with default [`MetadataOptionsEraseVersion`] options.
     ///
     /// Succeeds if the metadata does not exist.
     ///
     /// # Errors
     /// Returns a [`StorageError`] if there is an underlying store error.
     pub fn erase_metadata(&self) -> Result<(), StorageError> {
-        let storage_handle = Arc::new(StorageHandle::new(self.storage.clone()));
-        let storage_transformer = self
-            .storage_transformers()
-            .create_writable_transformer(storage_handle);
-        crate::storage::erase_metadata(&*storage_transformer, self.path())
+        self.erase_metadata_opt(&MetadataOptionsEraseVersion::default())
+    }
+
+    /// Erase the metadata with non-default [`MetadataOptionsEraseVersion`] options.
+    ///
+    /// Succeeds if the metadata does not exist.
+    ///
+    /// # Errors
+    /// Returns a [`StorageError`] if there is an underlying store error.
+    pub fn erase_metadata_opt(
+        &self,
+        options: &MetadataOptionsEraseVersion,
+    ) -> Result<(), StorageError> {
+        let storage_handle = StorageHandle::new(self.storage.clone());
+        match options {
+            MetadataOptionsEraseVersion::Default => match self.metadata {
+                ArrayMetadata::V3(_) => storage_handle.erase(&meta_key(self.path())),
+                ArrayMetadata::V2(_) => {
+                    storage_handle.erase(&meta_key_v2_array(self.path()))?;
+                    storage_handle.erase(&meta_key_v2_attributes(self.path()))
+                }
+            },
+            MetadataOptionsEraseVersion::All => {
+                storage_handle.erase(&meta_key(self.path()))?;
+                storage_handle.erase(&meta_key_v2_array(self.path()))?;
+                storage_handle.erase(&meta_key_v2_attributes(self.path()))
+            }
+            MetadataOptionsEraseVersion::V3 => storage_handle.erase(&meta_key(self.path())),
+            MetadataOptionsEraseVersion::V2 => {
+                storage_handle.erase(&meta_key_v2_array(self.path()))?;
+                storage_handle.erase(&meta_key_v2_attributes(self.path()))
+            }
+        }
     }
 
     /// Erase the chunk at `chunk_indices`.