+++ CHANGELOG.md
@@ -12,8 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Add `ArrayBuilder::build_arc()`
  - Add `Array::[async_]retrieve_encoded_chunk[s]`
  - Add internal `fill_array_view_with_fill_value` function
- - Add `Array::store_metadata_opt`
- - Add `ArrayMetadataOptionsVersion` and `{ArrayMetadataOptions,Config}::[set_]array_metadata_version`
+ - Add `{Array,Group}::{store,erase}_metadata_opt`
+ - Add `MetadataOptions{Store,Erase}Version`
+   - Add `{Config,{Array,Group}MetadataOptions}::[set_]metadata_{store,erase}_version`
 
 ### Changed
  - Support `object_store` 0.9-0.10